discord-websocket_way/requirements.txt
pycryptodome==3.15.0
discord.py==2.0.1
qrcode==7.3.1
aiohttp==3.8.3